Move source gen tests to the source gen test project.
